backend/requirements.txt

fastapi
uvicorn
python-multipart
pillow
requests
sarvamai
beautifulsoup4
python-dotenv